added random to special

+++ app/Http/Controllers/ProfileController.php
@@ -6,12 +6,16 @@ use App\Role;
 use App\Poste;
 use App\Testimonial;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
 
 use Illuminate\Http\Request;
 
 class ProfileController extends Controller
 {
     public function index(){
+        if(!Auth::check()){
+            return redirect()->back();
+        }
         $user=User::find(Auth::id());
         $role=Role::all()->where('id',$user->id_role)->first();
         $poste=Poste::all()->where('id',$user->id_poste)->first();
@@ -19,6 +23,9 @@ class ProfileController extends Controller
         return view('admin.profile.index',compact('user','role','poste','testimonial'));
     }
     public function edit(){
+        if(!Auth::check()){
+            return redirect()->back();
+        }
         $user=User::find(Auth::id());
         $roles=Role::all();
         $postes=Poste::all();
@@ -28,6 +35,9 @@ class ProfileController extends Controller
     }
 
     public function update(Request $request){
+        if(!Auth::check()){
+            return redirect()->back();
+        }
         $user=User::find(Auth::id());
         if(Storage::exists(public_path($user->photo))){
             unlink($user->photo);
@@ -39,14 +49,15 @@ class ProfileController extends Controller
         $user->prenom=$request->prenom;
         $user->photo=$file;
         $user->password=$request->password;
-        $user->id_role=$request->id_role;
-        $user->id_poste=$request->id_poste;
-        $user->id_testimonial=$request->id_testimonial;
+        // $user->id_testimonial=$request->id_testimonial;
         $user->save();
         return redirect()->route('profile');
 
     }
     public function destroy(){
+        if(!Auth::check()){
+            return redirect()->back();
+        }
         $user=User::find(Auth::id());
         if(Storage::exists(public_path($user->photo))){
             unlink($user->photo);
